+++ advanced_scripts/moving_text.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-# ffmpeg video slideshow script for advanced moving text v2 (01.12.2018)
+# ffmpeg video slideshow script for advanced moving text v4 (10.12.2018)
 #
 # Copyright (c) 2017-2018, Taner Sener (https://github.com/tanersener)
 #
@@ -13,6 +13,7 @@ HEIGHT=720
 FPS=30
 TRANSITION_DURATION=1
 PHOTO_DURATION=2
+PHOTO_MODE=3                # 1=CENTER, 2=CROP, 3=SCALE, 4=BLUR
 TEXT_FRAME_HEIGHT=60
 TEXT_FRAME_Y=660
 TEXT_Y=678
@@ -20,8 +21,9 @@ TEXT_FONT="../fonts/FallingSkyBd.otf"
 TEXT="5 Wonders of the World\:   1. Colosseum   2. The Great Pyramid & Sphinx Eygpt   3. Leaning Tower of Pisa   4. Taj Mahal   5. Chichen Itza"
 TEXT_FONT_SIZE=30
 TEXT_FONT_COLOR=white
-TEXT_SPEED=3            # 1=FASTEST, 2=FASTER, 3=MODERATE, 4=SLOW, 5=SLOWEST
+TEXT_SPEED=3                # 1=FASTEST, 2=FASTER, 3=MODERATE, 4=SLOW, 5=SLOWEST
 BACKGROUND_COLOR="#00000000"
+DIRECTION=2                 # 1=LEFT TO RIGHT, 2=RIGHT TO LEFT
 
 # PHOTO OPTIONS - ALL FILES UNDER photos FOLDER ARE USED - USE sort TO SPECIFY A SORTING MECHANISM
 # PHOTOS=`find ../photos/* | sort -r`
@@ -65,7 +67,22 @@ FULL_SCRIPT+="-filter_complex \""
 # 4. PREPARING SCALED INPUTS
 for (( c=0; c<${PHOTOS_COUNT}; c++ ))
 do
-    FULL_SCRIPT+="[${c}:v]setpts=PTS-STARTPTS,scale=w='if(gte(iw/ih,${WIDTH}/${HEIGHT}),min(iw,${WIDTH}),-1)':h='if(gte(iw/ih,${WIDTH}/${HEIGHT}),-1,min(ih,${HEIGHT}))',scale=trunc(iw/2)*2:trunc(ih/2)*2,setsar=sar=1/1,format=rgba,split=2[stream$((c+1))out1][stream$((c+1))out2];"
+    case ${PHOTO_MODE} in
+        1)
+            FULL_SCRIPT+="[${c}:v]setpts=PTS-STARTPTS,scale=w='if(gte(iw/ih,${WIDTH}/${HEIGHT}),min(iw,${WIDTH}),-1)':h='if(gte(iw/ih,${WIDTH}/${HEIGHT}),-1,min(ih,${HEIGHT}))',scale=trunc(iw/2)*2:trunc(ih/2)*2,setsar=sar=1/1,format=rgba,split=2[stream$((c+1))out1][stream$((c+1))out2];"
+        ;;
+        2)
+            FULL_SCRIPT+="[${c}:v]setpts=PTS-STARTPTS,scale=w='if(gte(iw/ih,${WIDTH}/${HEIGHT}),-1,${WIDTH})':h='if(gte(iw/ih,${WIDTH}/${HEIGHT}),${HEIGHT},-1)',crop=${WIDTH}:${HEIGHT},setsar=sar=1/1,format=rgba,split=2[stream$((c+1))out1][stream$((c+1))out2];"
+        ;;
+        3)
+            FULL_SCRIPT+="[${c}:v]setpts=PTS-STARTPTS,scale=${WIDTH}:${HEIGHT},setsar=sar=1/1,format=rgba,split=2[stream$((c+1))out1][stream$((c+1))out2];"
+        ;;
+        4)
+            FULL_SCRIPT+="[${c}:v]scale=${WIDTH}x${HEIGHT},setsar=sar=1/1,format=rgba,boxblur=100,setsar=sar=1/1[stream${c}blurred];"
+            FULL_SCRIPT+="[${c}:v]scale=w='if(gte(iw/ih,${WIDTH}/${HEIGHT}),min(iw,${WIDTH}),-1)':h='if(gte(iw/ih,${WIDTH}/${HEIGHT}),-1,min(ih,${HEIGHT}))',scale=trunc(iw/2)*2:trunc(ih/2)*2,setsar=sar=1/1,format=rgba[stream${c}raw];"
+            FULL_SCRIPT+="[stream${c}blurred][stream${c}raw]overlay=(main_w-overlay_w)/2:(main_h-overlay_h)/2:format=rgb,setpts=PTS-STARTPTS,split=2[stream$((c+1))out1][stream$((c+1))out2];"
+        ;;
+    esac
 done
 
 # 5. APPLYING PADDING
@@ -98,11 +115,18 @@ done
 # 8. END CONCAT
 FULL_SCRIPT+="[stream${PHOTOS_COUNT}overlaid]concat=n=$((2*PHOTOS_COUNT-1)):v=1:a=0,format=yuv420p[videowithouttext];"
 
-# 10. PREPARE HEART PHOTO INPUT
+# 10. PREPARE TEXT BOX
 FULL_SCRIPT+="[videowithouttext]drawbox=x=0:y=${TEXT_FRAME_Y}:w=${WIDTH}:h=${TEXT_FRAME_HEIGHT}:color=black@0.65:t=${TEXT_FRAME_HEIGHT},trim=duration=${TOTAL_DURATION}[videowithbox];"
 
 # 11. OVERLAY TEXT ON TOP OF SLIDESHOW
-FULL_SCRIPT+="[videowithbox]drawtext=fontfile=${TEXT_FONT}:text='${TEXT}':fontsize=${TEXT_FONT_SIZE}:fontcolor=${TEXT_FONT_COLOR}:x='${WIDTH} - mod(t/${TEXT_SPEED}*((${WIDTH}+text_w)/$((TEXT_SPEED))),${WIDTH}+text_w)':y=${TEXT_Y}[video]\""
+case ${DIRECTION} in
+    1)
+        FULL_SCRIPT+="[videowithbox]drawtext=fontfile=${TEXT_FONT}:text='${TEXT}':fontsize=${TEXT_FONT_SIZE}:fontcolor=${TEXT_FONT_COLOR}:x='-text_w + mod(t/${TEXT_SPEED}*((${WIDTH}+text_w)/$((TEXT_SPEED))),${WIDTH}+text_w)':y=${TEXT_Y}[video]\""
+    ;;
+    *)
+        FULL_SCRIPT+="[videowithbox]drawtext=fontfile=${TEXT_FONT}:text='${TEXT}':fontsize=${TEXT_FONT_SIZE}:fontcolor=${TEXT_FONT_COLOR}:x='${WIDTH} - mod(t/${TEXT_SPEED}*((${WIDTH}+text_w)/$((TEXT_SPEED))),${WIDTH}+text_w)':y=${TEXT_Y}[video]\""
+    ;;
+esac
 
 # 12. END
 FULL_SCRIPT+=" -map [video] -vsync 2 -async 1 -rc-lookahead 0 -g 0 -profile:v main -level 42 -c:v libx264 -r ${FPS} ../advanced_moving_text.mp4"