+++ transition_scripts/box_in_vertical.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-# ffmpeg video slideshow script with vertical box in transition v2 (04.12.2018)
+# ffmpeg video slideshow script with vertical box in transition v3 (10.12.2018)
 #
 # Copyright (c) 2017-2018, Taner Sener (https://github.com/tanersener)
 #
@@ -13,7 +13,9 @@ HEIGHT=720
 FPS=30
 TRANSITION_DURATION=1
 PHOTO_DURATION=2
+PHOTO_MODE=2                # 1=CENTER, 2=CROP, 3=SCALE, 4=BLUR
 BACKGROUND_COLOR="#00000000"
+DIRECTION=1                 # 1=TOP TO BOTTOM, 2=BOTTOM TO TOP
 
 # PHOTO OPTIONS - ALL FILES UNDER photos FOLDER ARE USED - USE sort TO SPECIFY A SORTING MECHANISM
 # PHOTOS=`find ../photos/* | sort -r`
@@ -61,7 +63,7 @@ for photo in ${PHOTOS}; do
     FULL_SCRIPT+="-loop 1 -i ${photo} "
 done
 
-# 3. ADD BACKGROUND COLOR INPUT
+# 3. ADD BACKGROUND COLOR SCREEN INPUT
 FULL_SCRIPT+="-f lavfi -i color=${BACKGROUND_COLOR}:s=${WIDTH}x${HEIGHT} "
 
 # 4. START FILTER COMPLEX
@@ -70,7 +72,22 @@ FULL_SCRIPT+="-filter_complex \""
 # 5. PREPARING SCALED INPUTS
 for (( c=0; c<${PHOTOS_COUNT}; c++ ))
 do
-    FULL_SCRIPT+="[${c}:v]setpts=PTS-STARTPTS,scale=w='if(gte(iw/ih,${WIDTH}/${HEIGHT}),min(iw,${WIDTH}),-1)':h='if(gte(iw/ih,${WIDTH}/${HEIGHT}),-1,min(ih,${HEIGHT}))',scale=trunc(iw/2)*2:trunc(ih/2)*2,pad=width=${WIDTH}:height=${HEIGHT}:x=(ow-iw)/2:y=(oh-ih)/2:color=${BACKGROUND_COLOR},setsar=sar=1/1,format=rgba,split=2[stream$((c+1))out1][stream$((c+1))out2];"
+    case ${PHOTO_MODE} in
+        1)
+            FULL_SCRIPT+="[${c}:v]setpts=PTS-STARTPTS,scale=w='if(gte(iw/ih,${WIDTH}/${HEIGHT}),min(iw,${WIDTH}),-1)':h='if(gte(iw/ih,${WIDTH}/${HEIGHT}),-1,min(ih,${HEIGHT}))',scale=trunc(iw/2)*2:trunc(ih/2)*2,pad=width=${WIDTH}:height=${HEIGHT}:x=(ow-iw)/2:y=(oh-ih)/2:color=${BACKGROUND_COLOR},setsar=sar=1/1,format=rgba,split=2[stream$((c+1))out1][stream$((c+1))out2];"
+        ;;
+        2)
+            FULL_SCRIPT+="[${c}:v]setpts=PTS-STARTPTS,scale=w='if(gte(iw/ih,${WIDTH}/${HEIGHT}),-1,${WIDTH})':h='if(gte(iw/ih,${WIDTH}/${HEIGHT}),${HEIGHT},-1)',crop=${WIDTH}:${HEIGHT},setsar=sar=1/1,format=rgba,split=2[stream$((c+1))out1][stream$((c+1))out2];"
+        ;;
+        3)
+            FULL_SCRIPT+="[${c}:v]setpts=PTS-STARTPTS,scale=w='if(gte(iw/ih,${WIDTH}/${HEIGHT}),min(iw,${WIDTH}),-1)':h='if(gte(iw/ih,${WIDTH}/${HEIGHT}),-1,min(ih,${HEIGHT}))',scale=${WIDTH}:${HEIGHT},setsar=sar=1/1,format=rgba,split=2[stream$((c+1))out1][stream$((c+1))out2];"
+        ;;
+        4)
+            FULL_SCRIPT+="[${c}:v]scale=${WIDTH}x${HEIGHT},setsar=sar=1/1,format=rgba,boxblur=100,setsar=sar=1/1[stream${c}blurred];"
+            FULL_SCRIPT+="[${c}:v]scale=w='if(gte(iw/ih,${WIDTH}/${HEIGHT}),min(iw,${WIDTH}),-1)':h='if(gte(iw/ih,${WIDTH}/${HEIGHT}),-1,min(ih,${HEIGHT}))',scale=trunc(iw/2)*2:trunc(ih/2)*2,setsar=sar=1/1,format=rgba[stream${c}raw];"
+            FULL_SCRIPT+="[stream${c}blurred][stream${c}raw]overlay=(main_w-overlay_w)/2:(main_h-overlay_h)/2:format=rgb,setpts=PTS-STARTPTS,split=2[stream$((c+1))out1][stream$((c+1))out2];"
+        ;;
+    esac
 
     FULL_SCRIPT+="[stream$((c+1))out1]trim=duration=${PHOTO_DURATION},select=lte(n\,${PHOTO_FRAME_COUNT})[stream$((c+1))overlaid];"
 
@@ -81,8 +98,16 @@ done
 # 6. OVERLAY INPUTS ON TOP OF BACKGROUND COLOR SCREEN
 for (( c=1; c<=${PHOTOS_COUNT}; c++ ))
 do
-    FULL_SCRIPT+="[${PHOTOS_COUNT}:v][stream${c}prephasein]overlay=x=0:y='${HEIGHT}-${HEIGHT}*t/(${TRANSITION_DURATION}/2)',trim=duration=${TRANSITION_PHASE_DURATION},select=lte(n\,(${TRANSITION_FRAME_COUNT}/2))[stream${c}phasein];"
-    FULL_SCRIPT+="[${PHOTOS_COUNT}:v][stream${c}prephaseout]overlay=x=0:y='${HEIGHT}*t/(${TRANSITION_DURATION}/2)',trim=duration=${TRANSITION_PHASE_DURATION},select=lte(n\,(${TRANSITION_FRAME_COUNT}/2))[stream${c}phaseout];"
+    case ${DIRECTION} in
+        1)
+            FULL_SCRIPT+="[${PHOTOS_COUNT}:v][stream${c}prephasein]overlay=x=0:y='-${HEIGHT}+${HEIGHT}*t/(${TRANSITION_DURATION}/2)',trim=duration=${TRANSITION_PHASE_DURATION},select=lte(n\,(${TRANSITION_FRAME_COUNT}/2))[stream${c}phasein];"
+            FULL_SCRIPT+="[${PHOTOS_COUNT}:v][stream${c}prephaseout]overlay=x=0:y='-${HEIGHT}*t/(${TRANSITION_DURATION}/2)',trim=duration=${TRANSITION_PHASE_DURATION},select=lte(n\,(${TRANSITION_FRAME_COUNT}/2))[stream${c}phaseout];"
+        ;;
+        *)
+            FULL_SCRIPT+="[${PHOTOS_COUNT}:v][stream${c}prephasein]overlay=x=0:y='${HEIGHT}-${HEIGHT}*t/(${TRANSITION_DURATION}/2)',trim=duration=${TRANSITION_PHASE_DURATION},select=lte(n\,(${TRANSITION_FRAME_COUNT}/2))[stream${c}phasein];"
+            FULL_SCRIPT+="[${PHOTOS_COUNT}:v][stream${c}prephaseout]overlay=x=0:y='${HEIGHT}*t/(${TRANSITION_DURATION}/2)',trim=duration=${TRANSITION_PHASE_DURATION},select=lte(n\,(${TRANSITION_FRAME_COUNT}/2))[stream${c}phaseout];"
+        ;;
+    esac
 
     FULL_SCRIPT+="[stream${c}checkpoint]trim=duration=${CHECKPOINT_DURATION},split=2[stream${c}checkin][stream${c}checkout];"
 done