+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## - 2018-12-10
+- Push box moved under advanced scripts in v2.x
+- Direction parameter implemented in v2.x
+
+## - 2018-12-09
+- Photo mode parameter implemented in v2.x
+
 ## - 2018-12-06
 - v1.x and v2.x branch documentations updated.